+++ src/main/java/com/sparta/realtomatoapp/common/ModifiedAuditingEntity.java
@@ -1,7 +1,6 @@
-package com.sparta.realtomatoapp.domain;
+package com.sparta.realtomatoapp.common;
 
 import jakarta.persistence.*;
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
@@ -9,14 +8,15 @@ import java.time.LocalDateTime;
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-public abstract class BaseAuditingEntity {
-    @CreatedDate
-    @Column(updatable = false)
-    @Temporal(TemporalType.TIMESTAMP)
-    private LocalDateTime createdAt;
+public abstract class ModifiedAuditingEntity extends BaseAuditingEntity {
 
     @LastModifiedDate
     @Column
     @Temporal(TemporalType.TIMESTAMP)
     private LocalDateTime modifiedAt;
+
+
+    public ModifiedAuditingEntity() {
+        super();
+    }
 }
